[ticket/15934] Move viewforum to a controller
[skip ci]

PHPBB3-15934
