+++ phpBB/includes/functions.php
@@ -2524,101 +2524,6 @@ function login_box($redirect = '', $l_explain = '', $l_success = '', $admin = fa
 	page_footer();
 }
 
-/**
-* Generate forum login box
-*/
-function login_forum_box($forum_data)
-{
-	global $db, $phpbb_container, $request, $template, $user, $phpbb_dispatcher, $phpbb_root_path, $phpEx;
-
-	$password = $request->variable('password', '', true);
-
-	$sql = 'SELECT forum_id
-		FROM ' . FORUMS_ACCESS_TABLE . '
-		WHERE forum_id = ' . $forum_data['forum_id'] . '
-			AND user_id = ' . $user->data['user_id'] . "
-			AND session_id = '" . $db->sql_escape($user->session_id) . "'";
-	$result = $db->sql_query($sql);
-	$row = $db->sql_fetchrow($result);
-	$db->sql_freeresult($result);
-
-	if ($row)
-	{
-		return true;
-	}
-
-	if ($password)
-	{
-		// Remove expired authorised sessions
-		$sql = 'SELECT f.session_id
-			FROM ' . FORUMS_ACCESS_TABLE . ' f
-			LEFT JOIN ' . SESSIONS_TABLE . ' s ON (f.session_id = s.session_id)
-			WHERE s.session_id IS NULL';
-		$result = $db->sql_query($sql);
-
-		if ($row = $db->sql_fetchrow($result))
-		{
-			$sql_in = array();
-			do
-			{
-				$sql_in[] = (string) $row['session_id'];
-			}
-			while ($row = $db->sql_fetchrow($result));
-
-			// Remove expired sessions
-			$sql = 'DELETE FROM ' . FORUMS_ACCESS_TABLE . '
-				WHERE ' . $db->sql_in_set('session_id', $sql_in);
-			$db->sql_query($sql);
-		}
-		$db->sql_freeresult($result);
-
-		/* @var $passwords_manager \phpbb\passwords\manager */
-		$passwords_manager = $phpbb_container->get('passwords.manager');
-
-		if ($passwords_manager->check($password, $forum_data['forum_password']))
-		{
-			$sql_ary = array(
-				'forum_id'		=> (int) $forum_data['forum_id'],
-				'user_id'		=> (int) $user->data['user_id'],
-				'session_id'	=> (string) $user->session_id,
-			);
-
-			$db->sql_query('INSERT INTO ' . FORUMS_ACCESS_TABLE . ' ' . $db->sql_build_array('INSERT', $sql_ary));
-
-			return true;
-		}
-
-		$template->assign_var('LOGIN_ERROR', $user->lang['WRONG_PASSWORD']);
-	}
-
-	/**
-	* Performing additional actions, load additional data on forum login
-	*
-	* @event core.login_forum_box
-	* @var	array	forum_data		Array with forum data
-	* @var	string	password		Password entered
-	* @since 3.1.0-RC3
-	*/
-	$vars = array('forum_data', 'password');
-	extract($phpbb_dispatcher->trigger_event('core.login_forum_box', compact($vars)));
-
-	page_header($user->lang['LOGIN']);
-
-	$template->assign_vars(array(
-		'FORUM_NAME'			=> isset($forum_data['forum_name']) ? $forum_data['forum_name'] : '',
-		'S_LOGIN_ACTION'		=> build_url(array('f')),
-		'S_HIDDEN_FIELDS'		=> build_hidden_fields(array('f' => $forum_data['forum_id'])))
-	);
-
-	$template->set_filenames(array(
-		'body' => 'login_forum.html')
-	);
-
-	make_jumpbox(append_sid("{$phpbb_root_path}viewforum.$phpEx"), $forum_data['forum_id']);
-
-	page_footer();
-}
-
 // Little helpers
 
 /**
