+++ phpBB/includes/functions_display.php
@@ -20,877 +20,11 @@ if (!defined('IN_PHPBB'))
 }
 
 /**
-* Display Forums
-*/
-function display_forums($root_data = '', $display_moderators = true, $return_moderators = false)
-{
-	global $db, $auth, $user, $template;
-	global $phpbb_root_path, $phpEx, $config;
-	global $request, $phpbb_dispatcher, $phpbb_container;
-
-	$forum_rows = $subforums = $forum_ids = $forum_ids_moderator = $forum_moderators = $active_forum_ary = array();
-	$parent_id = $visible_forums = 0;
-
-	// Mark forums read?
-	$mark_read = $request->variable('mark', '');
-
-	if ($mark_read == 'all')
-	{
-		$mark_read = '';
-	}
-
-	if (!$root_data)
-	{
-		if ($mark_read == 'forums')
-		{
-			$mark_read = 'all';
-		}
-
-		$root_data = array('forum_id' => 0);
-		$sql_where = '';
-	}
-	else
-	{
-		$sql_where = 'left_id > ' . $root_data['left_id'] . ' AND left_id < ' . $root_data['right_id'];
-	}
-
-	// Handle marking everything read
-	if ($mark_read == 'all')
-	{
-		$redirect = build_url(array('mark', 'hash', 'mark_time'));
-		meta_refresh(3, $redirect);
-
-		if (check_link_hash($request->variable('hash', ''), 'global'))
-		{
-			markread('all', false, false, $request->variable('mark_time', 0));
-
-			if ($request->is_ajax())
-			{
-				// Tell the ajax script what language vars and URL need to be replaced
-				$data = array(
-					'NO_UNREAD_POSTS'	=> $user->lang['NO_UNREAD_POSTS'],
-					'UNREAD_POSTS'		=> $user->lang['UNREAD_POSTS'],
-					'U_MARK_FORUMS'		=> ($user->data['is_registered'] || $config['load_anon_lastread']) ? append_sid("{$phpbb_root_path}index.$phpEx", 'hash=' . generate_link_hash('global') . '&mark=forums&mark_time=' . time()) : '',
-					'MESSAGE_TITLE'		=> $user->lang['INFORMATION'],
-					'MESSAGE_TEXT'		=> $user->lang['FORUMS_MARKED']
-				);
-				$json_response = new \phpbb\json_response();
-				$json_response->send($data);
-			}
-
-			trigger_error(
-				$user->lang['FORUMS_MARKED'] . '<br /><br />' .
-				sprintf($user->lang['RETURN_INDEX'], '<a href="' . $redirect . '">', '</a>')
-			);
-		}
-		else
-		{
-			trigger_error(sprintf($user->lang['RETURN_PAGE'], '<a href="' . $redirect . '">', '</a>'));
-		}
-	}
-
-	// Display list of active topics for this category?
-	$show_active = (isset($root_data['forum_flags']) && ($root_data['forum_flags'] & FORUM_FLAG_ACTIVE_TOPICS)) ? true : false;
-
-	$sql_array = array(
-		'SELECT'	=> 'f.*',
-		'FROM'		=> array(
-			FORUMS_TABLE		=> 'f'
-		),
-		'LEFT_JOIN'	=> array(),
-	);
-
-	if ($config['load_db_lastread'] && $user->data['is_registered'])
-	{
-		$sql_array['LEFT_JOIN'][] = array('FROM' => array(FORUMS_TRACK_TABLE => 'ft'), 'ON' => 'ft.user_id = ' . $user->data['user_id'] . ' AND ft.forum_id = f.forum_id');
-		$sql_array['SELECT'] .= ', ft.mark_time';
-	}
-	else if ($config['load_anon_lastread'] || $user->data['is_registered'])
-	{
-		$tracking_topics = $request->variable($config['cookie_name'] . '_track', '', true, \phpbb\request\request_interface::COOKIE);
-		$tracking_topics = ($tracking_topics) ? tracking_unserialize($tracking_topics) : array();
-
-		if (!$user->data['is_registered'])
-		{
-			$user->data['user_lastmark'] = (isset($tracking_topics['l'])) ? (int) (base_convert($tracking_topics['l'], 36, 10) + $config['board_startdate']) : 0;
-		}
-	}
-
-	if ($show_active)
-	{
-		$sql_array['LEFT_JOIN'][] = array(
-			'FROM'	=> array(FORUMS_ACCESS_TABLE => 'fa'),
-			'ON'	=> "fa.forum_id = f.forum_id AND fa.session_id = '" . $db->sql_escape($user->session_id) . "'"
-		);
-
-		$sql_array['SELECT'] .= ', fa.user_id';
-	}
-
-	$sql_ary = array(
-		'SELECT'	=> $sql_array['SELECT'],
-		'FROM'		=> $sql_array['FROM'],
-		'LEFT_JOIN'	=> $sql_array['LEFT_JOIN'],
-
-		'WHERE'		=> $sql_where,
-
-		'ORDER_BY'	=> 'f.left_id',
-	);
-
-	/**
-	* Event to modify the SQL query before the forum data is queried
-	*
-	* @event core.display_forums_modify_sql
-	* @var	array	sql_ary		The SQL array to get the data of the forums
-	* @since 3.1.0-a1
-	*/
-	$vars = array('sql_ary');
-	extract($phpbb_dispatcher->trigger_event('core.display_forums_modify_sql', compact($vars)));
-
-	$sql = $db->sql_build_query('SELECT', $sql_ary);
-	$result = $db->sql_query($sql);
-
-	$forum_tracking_info = $valid_categories = array();
-	$branch_root_id = $root_data['forum_id'];
-
-	/* @var $phpbb_content_visibility \phpbb\content_visibility */
-	$phpbb_content_visibility = $phpbb_container->get('content.visibility');
-
-	while ($row = $db->sql_fetchrow($result))
-	{
-		/**
-		* Event to modify the data set of a forum
-		*
-		* This event is triggered once per forum
-		*
-		* @event core.display_forums_modify_row
-		* @var	int		branch_root_id	Last top-level forum
-		* @var	array	row				The data of the forum
-		* @since 3.1.0-a1
-		*/
-		$vars = array('branch_root_id', 'row');
-		extract($phpbb_dispatcher->trigger_event('core.display_forums_modify_row', compact($vars)));
-
-		$forum_id = $row['forum_id'];
-
-		// Mark forums read?
-		if ($mark_read == 'forums')
-		{
-			if ($auth->acl_get('f_list', $forum_id))
-			{
-				$forum_ids[] = $forum_id;
-			}
-
-			continue;
-		}
-
-		// Category with no members
-		if ($row['forum_type'] == FORUM_CAT && ($row['left_id'] + 1 == $row['right_id']))
-		{
-			continue;
-		}
-
-		// Skip branch
-		if (isset($right_id))
-		{
-			if ($row['left_id'] < $right_id)
-			{
-				continue;
-			}
-			unset($right_id);
-		}
-
-		if (!$auth->acl_get('f_list', $forum_id))
-		{
-			// if the user does not have permissions to list this forum, skip everything until next branch
-			$right_id = $row['right_id'];
-			continue;
-		}
-
-		if ($config['load_db_lastread'] && $user->data['is_registered'])
-		{
-			$forum_tracking_info[$forum_id] = (!empty($row['mark_time'])) ? $row['mark_time'] : $user->data['user_lastmark'];
-		}
-		else if ($config['load_anon_lastread'] || $user->data['is_registered'])
-		{
-			if (!$user->data['is_registered'])
-			{
-				$user->data['user_lastmark'] = (isset($tracking_topics['l'])) ? (int) (base_convert($tracking_topics['l'], 36, 10) + $config['board_startdate']) : 0;
-			}
-			$forum_tracking_info[$forum_id] = (isset($tracking_topics['f'][$forum_id])) ? (int) (base_convert($tracking_topics['f'][$forum_id], 36, 10) + $config['board_startdate']) : $user->data['user_lastmark'];
-		}
-
-		// Lets check whether there are unapproved topics/posts, so we can display an information to moderators
-		$row['forum_id_unapproved_topics'] = ($auth->acl_get('m_approve', $forum_id) && $row['forum_topics_unapproved']) ? $forum_id : 0;
-		$row['forum_id_unapproved_posts'] = ($auth->acl_get('m_approve', $forum_id) && $row['forum_posts_unapproved']) ? $forum_id : 0;
-		$row['forum_posts'] = $phpbb_content_visibility->get_count('forum_posts', $row, $forum_id);
-		$row['forum_topics'] = $phpbb_content_visibility->get_count('forum_topics', $row, $forum_id);
-
-		// Display active topics from this forum?
-		if ($show_active && $row['forum_type'] == FORUM_POST && $auth->acl_get('f_read', $forum_id) && ($row['forum_flags'] & FORUM_FLAG_ACTIVE_TOPICS))
-		{
-			if (!isset($active_forum_ary['forum_topics']))
-			{
-				$active_forum_ary['forum_topics'] = 0;
-			}
-
-			if (!isset($active_forum_ary['forum_posts']))
-			{
-				$active_forum_ary['forum_posts'] = 0;
-			}
-
-			$active_forum_ary['forum_id'][]		= $forum_id;
-			$active_forum_ary['enable_icons'][]	= $row['enable_icons'];
-			$active_forum_ary['forum_topics']	+= $row['forum_topics'];
-			$active_forum_ary['forum_posts']	+= $row['forum_posts'];
-
-			// If this is a passworded forum we do not show active topics from it if the user is not authorised to view it...
-			if ($row['forum_password'] && $row['user_id'] != $user->data['user_id'])
-			{
-				$active_forum_ary['exclude_forum_id'][] = $forum_id;
-			}
-		}
-
-		// Fill list of categories with forums
-		if (isset($forum_rows[$row['parent_id']]))
-		{
-			$valid_categories[$row['parent_id']] = true;
-		}
-
-		//
-		if ($row['parent_id'] == $root_data['forum_id'] || $row['parent_id'] == $branch_root_id)
-		{
-			if ($row['forum_type'] != FORUM_CAT)
-			{
-				$forum_ids_moderator[] = (int) $forum_id;
-			}
-
-			// Direct child of current branch
-			$parent_id = $forum_id;
-			$forum_rows[$forum_id] = $row;
-
-			if ($row['forum_type'] == FORUM_CAT && $row['parent_id'] == $root_data['forum_id'])
-			{
-				$branch_root_id = $forum_id;
-			}
-			$forum_rows[$parent_id]['forum_id_last_post'] = $row['forum_id'];
-			$forum_rows[$parent_id]['forum_password_last_post'] = $row['forum_password'];
-			$forum_rows[$parent_id]['orig_forum_last_post_time'] = $row['forum_last_post_time'];
-		}
-		else if ($row['forum_type'] != FORUM_CAT)
-		{
-			$subforums[$parent_id][$forum_id]['display'] = ($row['display_on_index']) ? true : false;
-			$subforums[$parent_id][$forum_id]['name'] = $row['forum_name'];
-			$subforums[$parent_id][$forum_id]['orig_forum_last_post_time'] = $row['forum_last_post_time'];
-			$subforums[$parent_id][$forum_id]['children'] = array();
-			$subforums[$parent_id][$forum_id]['type'] = $row['forum_type'];
-
-			if (isset($subforums[$parent_id][$row['parent_id']]) && !$row['display_on_index'])
-			{
-				$subforums[$parent_id][$row['parent_id']]['children'][] = $forum_id;
-			}
-
-			if (!$forum_rows[$parent_id]['forum_id_unapproved_topics'] && $row['forum_id_unapproved_topics'])
-			{
-				$forum_rows[$parent_id]['forum_id_unapproved_topics'] = $forum_id;
-			}
-
-			if (!$forum_rows[$parent_id]['forum_id_unapproved_posts'] && $row['forum_id_unapproved_posts'])
-			{
-				$forum_rows[$parent_id]['forum_id_unapproved_posts'] = $forum_id;
-			}
-
-			$forum_rows[$parent_id]['forum_topics'] += $row['forum_topics'];
-
-			// Do not list redirects in LINK Forums as Posts.
-			if ($row['forum_type'] != FORUM_LINK)
-			{
-				$forum_rows[$parent_id]['forum_posts'] += $row['forum_posts'];
-			}
-
-			if ($row['forum_last_post_time'] > $forum_rows[$parent_id]['forum_last_post_time'])
-			{
-				$forum_rows[$parent_id]['forum_last_post_id'] = $row['forum_last_post_id'];
-				$forum_rows[$parent_id]['forum_last_post_subject'] = $row['forum_last_post_subject'];
-				$forum_rows[$parent_id]['forum_last_post_time'] = $row['forum_last_post_time'];
-				$forum_rows[$parent_id]['forum_last_poster_id'] = $row['forum_last_poster_id'];
-				$forum_rows[$parent_id]['forum_last_poster_name'] = $row['forum_last_poster_name'];
-				$forum_rows[$parent_id]['forum_last_poster_colour'] = $row['forum_last_poster_colour'];
-				$forum_rows[$parent_id]['forum_id_last_post'] = $forum_id;
-				$forum_rows[$parent_id]['forum_password_last_post'] = $row['forum_password'];
-			}
-		}
-
-		/**
-		* Event to modify the forum rows data set
-		*
-		* This event is triggered once per forum
-		*
-		* @event core.display_forums_modify_forum_rows
-		* @var	array	forum_rows		Data array of all forums we display
-		* @var	array	subforums		Data array of all subforums we display
-		* @var	int		branch_root_id	Current top-level forum
-		* @var	int		parent_id		Current parent forum
-		* @var	array	row				The data of the forum
-		* @since 3.1.0-a1
-		*/
-		$vars = array('forum_rows', 'subforums', 'branch_root_id', 'parent_id', 'row');
-		extract($phpbb_dispatcher->trigger_event('core.display_forums_modify_forum_rows', compact($vars)));
-	}
-	$db->sql_freeresult($result);
-
-	// Handle marking posts
-	if ($mark_read == 'forums')
-	{
-		$redirect = build_url(array('mark', 'hash', 'mark_time'));
-		$token = $request->variable('hash', '');
-		if (check_link_hash($token, 'global'))
-		{
-			markread('topics', $forum_ids, false, $request->variable('mark_time', 0));
-			$message = sprintf($user->lang['RETURN_FORUM'], '<a href="' . $redirect . '">', '</a>');
-			meta_refresh(3, $redirect);
-
-			if ($request->is_ajax())
-			{
-				// Tell the ajax script what language vars and URL need to be replaced
-				$data = array(
-					'NO_UNREAD_POSTS'	=> $user->lang['NO_UNREAD_POSTS'],
-					'UNREAD_POSTS'		=> $user->lang['UNREAD_POSTS'],
-					'U_MARK_FORUMS'		=> ($user->data['is_registered'] || $config['load_anon_lastread']) ? append_sid("{$phpbb_root_path}viewforum.$phpEx", 'hash=' . generate_link_hash('global') . '&f=' . $root_data['forum_id'] . '&mark=forums&mark_time=' . time()) : '',
-					'MESSAGE_TITLE'		=> $user->lang['INFORMATION'],
-					'MESSAGE_TEXT'		=> $user->lang['FORUMS_MARKED']
-				);
-				$json_response = new \phpbb\json_response();
-				$json_response->send($data);
-			}
-
-			trigger_error($user->lang['FORUMS_MARKED'] . '<br /><br />' . $message);
-		}
-		else
-		{
-			$message = sprintf($user->lang['RETURN_PAGE'], '<a href="' . $redirect . '">', '</a>');
-			meta_refresh(3, $redirect);
-			trigger_error($message);
-		}
-
-	}
-
-	// Grab moderators ... if necessary
-	if ($display_moderators)
-	{
-		if ($return_moderators)
-		{
-			$forum_ids_moderator[] = $root_data['forum_id'];
-		}
-		get_moderators($forum_moderators, $forum_ids_moderator);
-	}
-
-	/**
-	* Event to perform additional actions before the forum list is being generated
-	*
-	* @event core.display_forums_before
-	* @var	array	active_forum_ary	Array with forum data to display active topics
-	* @var	bool	display_moderators	Flag indicating if we display forum moderators
-	* @var	array	forum_moderators	Array with forum moderators list
-	* @var	array	forum_rows			Data array of all forums we display
-	* @var	bool	return_moderators	Flag indicating if moderators list should be returned
-	* @var	array	root_data			Array with the root forum data
-	* @since 3.1.4-RC1
-	*/
-	$vars = array(
-		'active_forum_ary',
-		'display_moderators',
-		'forum_moderators',
-		'forum_rows',
-		'return_moderators',
-		'root_data',
-	);
-	extract($phpbb_dispatcher->trigger_event('core.display_forums_before', compact($vars)));
-
-	// Used to tell whatever we have to create a dummy category or not.
-	$last_catless = true;
-	foreach ($forum_rows as $row)
-	{
-		// Category
-		if ($row['parent_id'] == $root_data['forum_id'] && $row['forum_type'] == FORUM_CAT)
-		{
-			// Do not display categories without any forums to display
-			if (!isset($valid_categories[$row['forum_id']]))
-			{
-				continue;
-			}
-
-			$cat_row = array(
-				'S_IS_CAT'				=> true,
-				'FORUM_ID'				=> $row['forum_id'],
-				'FORUM_NAME'			=> $row['forum_name'],
-				'FORUM_DESC'			=> generate_text_for_display($row['forum_desc'], $row['forum_desc_uid'], $row['forum_desc_bitfield'], $row['forum_desc_options']),
-				'FORUM_FOLDER_IMG'		=> '',
-				'FORUM_FOLDER_IMG_SRC'	=> '',
-				'FORUM_IMAGE'			=> ($row['forum_image']) ? '<img src="' . $phpbb_root_path . $row['forum_image'] . '" alt="' . $user->lang['FORUM_CAT'] . '" />' : '',
-				'FORUM_IMAGE_SRC'		=> ($row['forum_image']) ? $phpbb_root_path . $row['forum_image'] : '',
-				'U_VIEWFORUM'			=> append_sid("{$phpbb_root_path}viewforum.$phpEx", 'f=' . $row['forum_id']),
-			);
-
-			/**
-			* Modify the template data block of the 'category'
-			*
-			* This event is triggered once per 'category'
-			*
-			* @event core.display_forums_modify_category_template_vars
-			* @var	array	cat_row			Template data of the 'category'
-			* @var	bool	last_catless	The flag indicating whether the last forum had a parent category
-			* @var	array	root_data		Array with the root forum data
-			* @var	array	row				The data of the 'category'
-			* @since 3.1.0-RC4
-			* @changed 3.1.7-RC1 Removed undefined catless variable
-			*/
-			$vars = array(
-				'cat_row',
-				'last_catless',
-				'root_data',
-				'row',
-			);
-			extract($phpbb_dispatcher->trigger_event('core.display_forums_modify_category_template_vars', compact($vars)));
-
-			$template->assign_block_vars('forumrow', $cat_row);
-
-			continue;
-		}
-
-		$visible_forums++;
-		$forum_id = $row['forum_id'];
-
-		$forum_unread = (isset($forum_tracking_info[$forum_id]) && $row['orig_forum_last_post_time'] > $forum_tracking_info[$forum_id]) ? true : false;
-
-		$folder_image = $folder_alt = $l_subforums = '';
-		$subforums_list = array();
-
-		// Generate list of subforums if we need to
-		if (isset($subforums[$forum_id]))
-		{
-			foreach ($subforums[$forum_id] as $subforum_id => $subforum_row)
-			{
-				$subforum_unread = (isset($forum_tracking_info[$subforum_id]) && $subforum_row['orig_forum_last_post_time'] > $forum_tracking_info[$subforum_id]) ? true : false;
-
-				if (!$subforum_unread && !empty($subforum_row['children']))
-				{
-					foreach ($subforum_row['children'] as $child_id)
-					{
-						if (isset($forum_tracking_info[$child_id]) && $subforums[$forum_id][$child_id]['orig_forum_last_post_time'] > $forum_tracking_info[$child_id])
-						{
-							// Once we found an unread child forum, we can drop out of this loop
-							$subforum_unread = true;
-							break;
-						}
-					}
-				}
-
-				if ($subforum_row['display'] && $subforum_row['name'])
-				{
-					$subforums_list[] = array(
-						'link'		=> append_sid("{$phpbb_root_path}viewforum.$phpEx", 'f=' . $subforum_id),
-						'name'		=> $subforum_row['name'],
-						'unread'	=> $subforum_unread,
-						'type'		=> $subforum_row['type'],
-					);
-				}
-				else
-				{
-					unset($subforums[$forum_id][$subforum_id]);
-				}
-
-				// If one subforum is unread the forum gets unread too...
-				if ($subforum_unread)
-				{
-					$forum_unread = true;
-				}
-			}
-
-			$l_subforums = (count($subforums[$forum_id]) == 1) ? $user->lang['SUBFORUM'] : $user->lang['SUBFORUMS'];
-			$folder_image = ($forum_unread) ? 'forum_unread_subforum' : 'forum_read_subforum';
-		}
-		else
-		{
-			switch ($row['forum_type'])
-			{
-				case FORUM_POST:
-					$folder_image = ($forum_unread) ? 'forum_unread' : 'forum_read';
-				break;
-
-				case FORUM_LINK:
-					$folder_image = 'forum_link';
-				break;
-			}
-		}
-
-		// Which folder should we display?
-		if ($row['forum_status'] == ITEM_LOCKED)
-		{
-			$folder_image = ($forum_unread) ? 'forum_unread_locked' : 'forum_read_locked';
-			$folder_alt = 'FORUM_LOCKED';
-		}
-		else
-		{
-			$folder_alt = ($forum_unread) ? 'UNREAD_POSTS' : 'NO_UNREAD_POSTS';
-		}
-
-		// Create last post link information, if appropriate
-		if ($row['forum_last_post_id'])
-		{
-			if ($row['forum_password_last_post'] === '' && $auth->acl_gets('f_read', 'f_list_topics', $row['forum_id_last_post']))
-			{
-				$last_post_subject = censor_text($row['forum_last_post_subject']);
-				$last_post_subject_truncated = truncate_string($last_post_subject, 30, 255, false, $user->lang['ELLIPSIS']);
-			}
-			else
-			{
-				$last_post_subject = $last_post_subject_truncated = '';
-			}
-			$last_post_time = $user->format_date($row['forum_last_post_time']);
-			$last_post_url = append_sid("{$phpbb_root_path}viewtopic.$phpEx", 'f=' . $row['forum_id_last_post'] . '&amp;p=' . $row['forum_last_post_id']) . '#p' . $row['forum_last_post_id'];
-		}
-		else
-		{
-			$last_post_subject = $last_post_time = $last_post_url = $last_post_subject_truncated = '';
-		}
-
-		// Output moderator listing ... if applicable
-		$l_moderator = $moderators_list = '';
-		if ($display_moderators && !empty($forum_moderators[$forum_id]))
-		{
-			$l_moderator = (count($forum_moderators[$forum_id]) == 1) ? $user->lang['MODERATOR'] : $user->lang['MODERATORS'];
-			$moderators_list = implode($user->lang['COMMA_SEPARATOR'], $forum_moderators[$forum_id]);
-		}
-
-		$l_post_click_count = ($row['forum_type'] == FORUM_LINK) ? 'CLICKS' : 'POSTS';
-		$post_click_count = ($row['forum_type'] != FORUM_LINK || $row['forum_flags'] & FORUM_FLAG_LINK_TRACK) ? $row['forum_posts'] : '';
-
-		$s_subforums_list = $subforums_row = array();
-		foreach ($subforums_list as $subforum)
-		{
-			$s_subforums_list[] = '<a href="' . $subforum['link'] . '" class="subforum ' . (($subforum['unread']) ? 'unread' : 'read') . '" title="' . (($subforum['unread']) ? $user->lang['UNREAD_POSTS'] : $user->lang['NO_UNREAD_POSTS']) . '">' . $subforum['name'] . '</a>';
-			$subforums_row[] = array(
-				'U_SUBFORUM'	=> $subforum['link'],
-				'SUBFORUM_NAME'	=> $subforum['name'],
-				'S_UNREAD'		=> $subforum['unread'],
-				'IS_LINK'		=> $subforum['type'] == FORUM_LINK,
-			);
-		}
-		$s_subforums_list = (string) implode($user->lang['COMMA_SEPARATOR'], $s_subforums_list);
-		$catless = ($row['parent_id'] == $root_data['forum_id']) ? true : false;
-
-		if ($row['forum_type'] != FORUM_LINK)
-		{
-			$u_viewforum = append_sid("{$phpbb_root_path}viewforum.$phpEx", 'f=' . $row['forum_id']);
-		}
-		else
-		{
-			// If the forum is a link and we count redirects we need to visit it
-			// If the forum is having a password or no read access we do not expose the link, but instead handle it in viewforum
-			if (($row['forum_flags'] & FORUM_FLAG_LINK_TRACK) || $row['forum_password'] || !$auth->acl_get('f_read', $forum_id))
-			{
-				$u_viewforum = append_sid("{$phpbb_root_path}viewforum.$phpEx", 'f=' . $row['forum_id']);
-			}
-			else
-			{
-				$u_viewforum = $row['forum_link'];
-			}
-		}
-
-		$forum_row = array(
-			'S_IS_CAT'			=> false,
-			'S_NO_CAT'			=> $catless && !$last_catless,
-			'S_IS_LINK'			=> ($row['forum_type'] == FORUM_LINK) ? true : false,
-			'S_UNREAD_FORUM'	=> $forum_unread,
-			'S_AUTH_READ'		=> $auth->acl_get('f_read', $row['forum_id']),
-			'S_LOCKED_FORUM'	=> ($row['forum_status'] == ITEM_LOCKED) ? true : false,
-			'S_LIST_SUBFORUMS'	=> ($row['display_subforum_list']) ? true : false,
-			'S_SUBFORUMS'		=> (count($subforums_list)) ? true : false,
-			'S_DISPLAY_SUBJECT'	=>	($last_post_subject !== '' && $config['display_last_subject']) ? true : false,
-			'S_FEED_ENABLED'	=> ($config['feed_forum'] && !phpbb_optionget(FORUM_OPTION_FEED_EXCLUDE, $row['forum_options']) && $row['forum_type'] == FORUM_POST) ? true : false,
-
-			'FORUM_ID'				=> $row['forum_id'],
-			'FORUM_NAME'			=> $row['forum_name'],
-			'FORUM_DESC'			=> generate_text_for_display($row['forum_desc'], $row['forum_desc_uid'], $row['forum_desc_bitfield'], $row['forum_desc_options']),
-			'TOPICS'				=> $row['forum_topics'],
-			$l_post_click_count		=> $post_click_count,
-			'FORUM_IMG_STYLE'		=> $folder_image,
-			'FORUM_FOLDER_IMG'		=> $user->img($folder_image, $folder_alt),
-			'FORUM_FOLDER_IMG_ALT'	=> isset($user->lang[$folder_alt]) ? $user->lang[$folder_alt] : '',
-			'FORUM_IMAGE'			=> ($row['forum_image']) ? '<img src="' . $phpbb_root_path . $row['forum_image'] . '" alt="' . $user->lang[$folder_alt] . '" />' : '',
-			'FORUM_IMAGE_SRC'		=> ($row['forum_image']) ? $phpbb_root_path . $row['forum_image'] : '',
-			'LAST_POST_SUBJECT'		=> $last_post_subject,
-			'LAST_POST_SUBJECT_TRUNCATED'	=> $last_post_subject_truncated,
-			'LAST_POST_TIME'		=> $last_post_time,
-			'LAST_POSTER'			=> get_username_string('username', $row['forum_last_poster_id'], $row['forum_last_poster_name'], $row['forum_last_poster_colour']),
-			'LAST_POSTER_COLOUR'	=> get_username_string('colour', $row['forum_last_poster_id'], $row['forum_last_poster_name'], $row['forum_last_poster_colour']),
-			'LAST_POSTER_FULL'		=> get_username_string('full', $row['forum_last_poster_id'], $row['forum_last_poster_name'], $row['forum_last_poster_colour']),
-			'MODERATORS'			=> $moderators_list,
-			'SUBFORUMS'				=> $s_subforums_list,
-
-			'L_SUBFORUM_STR'		=> $l_subforums,
-			'L_MODERATOR_STR'		=> $l_moderator,
-
-			'U_UNAPPROVED_TOPICS'	=> ($row['forum_id_unapproved_topics']) ? append_sid("{$phpbb_root_path}mcp.$phpEx", 'i=queue&amp;mode=unapproved_topics&amp;f=' . $row['forum_id_unapproved_topics']) : '',
-			'U_UNAPPROVED_POSTS'	=> ($row['forum_id_unapproved_posts']) ? append_sid("{$phpbb_root_path}mcp.$phpEx", 'i=queue&amp;mode=unapproved_posts&amp;f=' . $row['forum_id_unapproved_posts']) : '',
-			'U_VIEWFORUM'		=> $u_viewforum,
-			'U_LAST_POSTER'		=> get_username_string('profile', $row['forum_last_poster_id'], $row['forum_last_poster_name'], $row['forum_last_poster_colour']),
-			'U_LAST_POST'		=> $last_post_url,
-		);
-
-		/**
-		* Modify the template data block of the forum
-		*
-		* This event is triggered once per forum
-		*
-		* @event core.display_forums_modify_template_vars
-		* @var	array	forum_row		Template data of the forum
-		* @var	array	row				The data of the forum
-		* @var	array	subforums_row	Template data of subforums
-		* @since 3.1.0-a1
-		* @changed 3.1.0-b5 Added var subforums_row
-		*/
-		$vars = array('forum_row', 'row', 'subforums_row');
-		extract($phpbb_dispatcher->trigger_event('core.display_forums_modify_template_vars', compact($vars)));
-
-		$template->assign_block_vars('forumrow', $forum_row);
-
-		// Assign subforums loop for style authors
-		$template->assign_block_vars_array('forumrow.subforum', $subforums_row);
-
-		/**
-		* Modify and/or assign additional template data for the forum
-		* after forumrow loop has been assigned. This can be used
-		* to create additional forumrow subloops in extensions.
-		*
-		* This event is triggered once per forum
-		*
-		* @event core.display_forums_add_template_data
-		* @var	array	forum_row		Template data of the forum
-		* @var	array	row				The data of the forum
-		* @var	array	subforums_list	The data of subforums
-		* @var	array	subforums_row	Template data of subforums
-		* @var	bool	catless			The flag indicating whether a forum has a parent category
-		* @since 3.1.0-b5
-		*/
-		$vars = array(
-			'forum_row',
-			'row',
-			'subforums_list',
-			'subforums_row',
-			'catless',
-		);
-		extract($phpbb_dispatcher->trigger_event('core.display_forums_add_template_data', compact($vars)));
-
-		$last_catless = $catless;
-	}
-
-	$template->assign_vars(array(
-		'U_MARK_FORUMS'		=> ($user->data['is_registered'] || $config['load_anon_lastread']) ? append_sid("{$phpbb_root_path}viewforum.$phpEx", 'hash=' . generate_link_hash('global') . '&amp;f=' . $root_data['forum_id'] . '&amp;mark=forums&amp;mark_time=' . time()) : '',
-		'S_HAS_SUBFORUM'	=> ($visible_forums) ? true : false,
-		'L_SUBFORUM'		=> ($visible_forums == 1) ? $user->lang['SUBFORUM'] : $user->lang['SUBFORUMS'],
-		'LAST_POST_IMG'		=> $user->img('icon_topic_latest', 'VIEW_LATEST_POST'),
-		'UNAPPROVED_IMG'	=> $user->img('icon_topic_unapproved', 'TOPICS_UNAPPROVED'),
-		'UNAPPROVED_POST_IMG'	=> $user->img('icon_topic_unapproved', 'POSTS_UNAPPROVED_FORUM'),
-	));
-
-	/**
-	* Event to perform additional actions after the forum list has been generated
-	*
-	* @event core.display_forums_after
-	* @var	array	active_forum_ary	Array with forum data to display active topics
-	* @var	bool	display_moderators	Flag indicating if we display forum moderators
-	* @var	array	forum_moderators	Array with forum moderators list
-	* @var	array	forum_rows			Data array of all forums we display
-	* @var	bool	return_moderators	Flag indicating if moderators list should be returned
-	* @var	array	root_data			Array with the root forum data
-	* @since 3.1.0-RC5
-	*/
-	$vars = array(
-		'active_forum_ary',
-		'display_moderators',
-		'forum_moderators',
-		'forum_rows',
-		'return_moderators',
-		'root_data',
-	);
-	extract($phpbb_dispatcher->trigger_event('core.display_forums_after', compact($vars)));
-
-	if ($return_moderators)
-	{
-		return array($active_forum_ary, $forum_moderators);
-	}
-
-	return array($active_forum_ary, array());
-}
-
-/**
-* Create forum rules for given forum
-*/
-function generate_forum_rules(&$forum_data)
-{
-	if ($forum_data['forum_rules'])
-	{
-		$forum_data['forum_rules'] = generate_text_for_display($forum_data['forum_rules'], $forum_data['forum_rules_uid'], $forum_data['forum_rules_bitfield'], $forum_data['forum_rules_options']);
-	}
-
-	if (!$forum_data['forum_rules'] && !$forum_data['forum_rules_link'])
-	{
-		return;
-	}
-
-	global $template;
-
-	$template->assign_vars(array(
-		'S_FORUM_RULES'	=> true,
-		'U_FORUM_RULES'	=> $forum_data['forum_rules_link'],
-		'FORUM_RULES'	=> $forum_data['forum_rules'])
-	);
-}
-
-/**
-* Create forum navigation links for given forum, create parent
-* list if currently null, assign basic forum info to template
-*/
-function generate_forum_nav(&$forum_data_ary)
-{
-	global $template, $auth, $config;
-	global $phpEx, $phpbb_root_path, $phpbb_dispatcher;
-
-	if (!$auth->acl_get('f_list', $forum_data_ary['forum_id']))
-	{
-		return;
-	}
-
-	$navlinks_parents = $forum_template_data = array();
-
-	// Get forum parents
-	$forum_parents = get_forum_parents($forum_data_ary);
-
-	$microdata_attr = 'data-forum-id';
-
-	// Build navigation links
-	if (!empty($forum_parents))
-	{
-		foreach ($forum_parents as $parent_forum_id => $parent_data)
-		{
-			list($parent_name, $parent_type) = array_values($parent_data);
-
-			// Skip this parent if the user does not have the permission to view it
-			if (!$auth->acl_get('f_list', $parent_forum_id))
-			{
-				continue;
-			}
-
-			$navlinks_parents[] = array(
-				'S_IS_CAT'			=> ($parent_type == FORUM_CAT) ? true : false,
-				'S_IS_LINK'			=> ($parent_type == FORUM_LINK) ? true : false,
-				'S_IS_POST'			=> ($parent_type == FORUM_POST) ? true : false,
-				'BREADCRUMB_NAME'	=> $parent_name,
-				'FORUM_ID'			=> $parent_forum_id,
-				'MICRODATA'			=> $microdata_attr . '="' . $parent_forum_id . '"',
-				'U_BREADCRUMB'		=> append_sid("{$phpbb_root_path}viewforum.$phpEx", 'f=' . $parent_forum_id),
-			);
-		}
-	}
-
-	$navlinks = array(
-		'S_IS_CAT'			=> ($forum_data_ary['forum_type'] == FORUM_CAT) ? true : false,
-		'S_IS_LINK'			=> ($forum_data_ary['forum_type'] == FORUM_LINK) ? true : false,
-		'S_IS_POST'			=> ($forum_data_ary['forum_type'] == FORUM_POST) ? true : false,
-		'BREADCRUMB_NAME'	=> $forum_data_ary['forum_name'],
-		'FORUM_ID'			=> $forum_data_ary['forum_id'],
-		'MICRODATA'			=> $microdata_attr . '="' . $forum_data_ary['forum_id'] . '"',
-		'U_BREADCRUMB'		=> append_sid("{$phpbb_root_path}viewforum.$phpEx", 'f=' . $forum_data_ary['forum_id']),
-	);
-
-	$forum_template_data = array(
-		'FORUM_ID' 		=> $forum_data_ary['forum_id'],
-		'FORUM_NAME'	=> $forum_data_ary['forum_name'],
-		'FORUM_DESC'	=> generate_text_for_display($forum_data_ary['forum_desc'], $forum_data_ary['forum_desc_uid'], $forum_data_ary['forum_desc_bitfield'], $forum_data_ary['forum_desc_options']),
-
-		'S_ENABLE_FEEDS_FORUM'	=> ($config['feed_forum'] && $forum_data_ary['forum_type'] == FORUM_POST && !phpbb_optionget(FORUM_OPTION_FEED_EXCLUDE, $forum_data_ary['forum_options'])) ? true : false,
-	);
-
-	$forum_data = $forum_data_ary;
-	/**
-	* Event to modify the navlinks text
-	*
-	* @event core.generate_forum_nav
-	* @var	array	forum_data				Array with the forum data
-	* @var	array	forum_template_data		Array with generic forum template data
-	* @var	string	microdata_attr			The microdata attribute
-	* @var	array	navlinks_parents		Array with the forum parents navlinks data
-	* @var	array	navlinks				Array with the forum navlinks data
-	* @since 3.1.5-RC1
-	*/
-	$vars = array(
-		'forum_data',
-		'forum_template_data',
-		'microdata_attr',
-		'navlinks_parents',
-		'navlinks',
-	);
-	extract($phpbb_dispatcher->trigger_event('core.generate_forum_nav', compact($vars)));
-	$forum_data_ary = $forum_data;
-	unset($forum_data);
-
-	$template->assign_block_vars_array('navlinks', $navlinks_parents);
-	$template->assign_block_vars('navlinks', $navlinks);
-	$template->assign_vars($forum_template_data);
-
-	return;
-}
-
-/**
-* Returns forum parents as an array. Get them from forum_data if available, or update the database otherwise
-*/
-function get_forum_parents(&$forum_data)
-{
-	global $db;
-
-	$forum_parents = array();
-
-	if ($forum_data['parent_id'] > 0)
-	{
-		if ($forum_data['forum_parents'] == '')
-		{
-			$sql = 'SELECT forum_id, forum_name, forum_type
-				FROM ' . FORUMS_TABLE . '
-				WHERE left_id < ' . $forum_data['left_id'] . '
-					AND right_id > ' . $forum_data['right_id'] . '
-				ORDER BY left_id ASC';
-			$result = $db->sql_query($sql);
-
-			while ($row = $db->sql_fetchrow($result))
-			{
-				$forum_parents[$row['forum_id']] = array($row['forum_name'], (int) $row['forum_type']);
-			}
-			$db->sql_freeresult($result);
-
-			$forum_data['forum_parents'] = serialize($forum_parents);
-
-			$sql = 'UPDATE ' . FORUMS_TABLE . "
-				SET forum_parents = '" . $db->sql_escape($forum_data['forum_parents']) . "'
-				WHERE parent_id = " . $forum_data['parent_id'];
-			$db->sql_query($sql);
-		}
-		else
-		{
-			$forum_parents = unserialize($forum_data['forum_parents']);
-		}
-	}
-
-	return $forum_parents;
-}
-
-/**
 * Obtain list of moderators of each forum
 */
 function get_moderators(&$forum_moderators, $forum_id = false)
 {
-	global $db, $phpbb_root_path, $phpEx, $user, $auth;
-	global $phpbb_container;
+	global $db, $phpbb_root_path, $phpEx, $user, $auth, $phpbb_container;
 
 	$forum_id_ary = array();
