+++ phpBB/styles/prosilver/template/forumlist_body.html
@@ -42,15 +42,19 @@
 								</a>
 							-->
 						<!-- ENDIF -->
-						<!-- IF forumrow.FORUM_IMAGE -->
+						<!-- IF forumrow.FORUM_IMAGE_SRC -->
 							<!-- EVENT forumlist_body_forum_image_before -->
-							<span class="forum-image"><!-- EVENT forumlist_body_forum_image_prepend -->{forumrow.FORUM_IMAGE}<!-- EVENT forumlist_body_forum_image_append --></span>
+							<span class="forum-image"><!-- EVENT forumlist_body_forum_image_prepend --><img src="{forumrow.FORUM_IMAGE_SRC}" alt="{forumrow.FORUM_IMAGE_ALT}"><!-- EVENT forumlist_body_forum_image_append --></span>
 							<!-- EVENT forumlist_body_forum_image_after -->
 						<!-- ENDIF -->
 						<a href="{forumrow.U_VIEWFORUM}" class="forumtitle">{forumrow.FORUM_NAME}</a>
 						<!-- IF forumrow.FORUM_DESC --><span class="forum-desc">{forumrow.FORUM_DESC}</span><!-- ENDIF -->
 						<!-- IF forumrow.MODERATORS -->
-							<span class="forum-mods"><strong>{forumrow.L_MODERATOR_STR}{L_COLON}</strong> {forumrow.MODERATORS}</span>
+							<span class="forum-mods"><strong>{forumrow.L_MODERATOR_STR}{L_COLON}</strong>
+							{% for moderator in forumrow.MODERATORS %}
+								{{ moderator }}{% if not loop.last %}{{ lang('COMMA_SEPARATOR') }}{% endif %}
+							{% endfor %}
+							</span>
 						<!-- ENDIF -->
 						<!-- IF .forumrow.subforum and forumrow.S_LIST_SUBFORUMS -->
 							<!-- EVENT forumlist_body_subforums_before -->
